Create An Analysis on Maternal Mental Health & Infant Sleep Patterns in Python
This project aimed to apply the skills I learned in my Predictive Analytics with Python course to a large data set. I have worked in healthcare for many years, especially with women. I found the 'Impact of Maternal Mental Health on Infant Sleep' dataset on kaggle.com interesting because of my own experience with being a mother and giving birth. I wanted to see if there was a connection between maternal mental health and infant sleep patterns since so many women suffer from post-partum depression. 
At the time the data was accessed Feb 6, 2023, it was only 17 days old and there was no other analysis done on kaggle.com working with this particular dataset. The set has 64 columns and 410 rows. Although there was no other analysis done on the ‘maternal mental health and infant sleep patterns’  dataset I was able to find Python analysis done on data sets concerning mental health. There was another analysis done on a different dataset concerning the mental health of Tech workers which was found on kaggle.com. 
I wanted to see if there was a correlation between poor maternal mental health and infant sleep patterns within the first three months of a child’s life. I decided to focus on the columns initially named [Age],[Gestationnal_age],[Type_pregnancy],[CBTS_M_3], [EPDS3][IBQ_R_VSF_3bb1], and [Sleep_night_duration]. While cleaning the data these columns were renamed ‘MomsAge’, ‘WeeksPregant’, ‘CityBirthTraumaSurvey3months’, ‘DepressionSurvey3months’, ‘Anxiety3months’, ‘InfantBehaviorSurvey3months’, and ‘BabySleep_duration’.  
Decided to use the Coefficient of Determination(R squared) to see if there is any relationship between the columns. I was able to do a heat map in Seaborn for all of the columns except for 'infantBehavior3months' and 'BabySleep_duration'. The columns with the highest cm2 of 0.10 were 'DepressionSurvey3months' and 'Anxiety3months', the 2nd highest cm2 of 0.05 was 'CityBirthTrumaScale3months' & 'Anxiety3months', the 3rd highest cm3 of 0.03. I was not able to find if columns 'infantbehavior3months' or 'BabySleep_duration' had a relation to any of the other columns because these columns were not mapped. I was not able to find a strong correlation between any of the columns with the heatmap. 
I attempted to use scatter, residual, and linear regression plots for the ‘Maternal Mental Health and infant sleep Patterns’ dataset however, the graphs produced were incompatible with the data. I then decided to use histoplots and boxplots which displayed little or no correlation between the columns as well.  
In conclusion, there was little or low correlation found between the columns analyzed. The R squared heatmap displayed low values, the seaborn linear regression plots were incompatible with the ‘Maternal mental health and infant sleep patterns’ data set, plus the boxplots and histoplots exhibit little relationship between columns. This may be due to the fact that this is test data or this data set was meant for a specific usage not mentioned in the description on the website from which it was accessed.